LPS-194841 Inline class AssetRendererSharingEntryInterpreter in class SharingEntryInterpreterProviderImpl
